+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.0
+
+- check git repository
+- optimize help message
+
 # 0.0.0
 
 first version